refactoring home page (#23)
