+++ Client/src/layouts/MainLayout.jsx
@@ -2,8 +2,7 @@
 import { Outlet } from "react-router-dom";
 import {ToastContainer} from "react-toastify";
 import Footer from "../modules/core/components/Footer.jsx";
-import Navbar from "../modules/core/system-component/Navbar.jsx";
-import SliderBanner from "../modules/core/components/SliderBanner.jsx";
+import Navbar from "../modules/core/components/Navbar.jsx";
 
 export function MainLayout() {
     return (
@@ -11,11 +10,10 @@ export function MainLayout() {
 
             <Navbar />
 
-      <div className={"mt-24"}>
+      <div>
         <div className={"w-full"}></div>
         <Outlet />
       </div>
-      <SliderBanner />
       <ToastContainer />
       <Footer />
     </div>
